implementing identity management

+++ ContosoPets.Ui/Pages/Products/Index.cshtml.cs
@@ -1,32 +1,42 @@
-﻿using ContosoPets.Ui.Extensions;
+using ContosoPets.Ui.Extensions;
 using ContosoPets.Ui.Models;
 using ContosoPets.Ui.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using System.Linq;
+using Microsoft.AspNetCore.Authorization;
 
 namespace ContosoPets.Ui.Pages.Products
 {
+    // Add [Authorize] attribute
+    [Authorize]
     public class IndexModel : PageModel
     {
         private readonly ProductService _productService;
 
         public string AntiforgeryToken => HttpContext.GetAntiforgeryTokenForJs();
         public IEnumerable<Product> Products { get; private set; } = new List<Product>();
+        // Add IsAdmin property
+        public bool IsAdmin =>
+            HttpContext.User.HasClaim("IsAdmin", bool.TrueString);
 
         public IndexModel(ProductService productService)
         {
             _productService = productService;
         }
 
-        public async Task OnGet()
-        {
+        public async Task OnGet() =>
             Products = await _productService.GetProducts();
-        }
     
         public async Task<IActionResult> OnDelete(int productId)
         {
+            // Add IsAdmin check
+            if (!IsAdmin)
+            {
+                return Forbid();
+            }
             try
             {
                 await _productService.DeleteProduct(productId);
